Refactor the code to resolve codeclimate issues
